User Guide for adding the watermark to PDF
This is a guide I created for our Control Coordinator to be able to add the needed JavaScript to assist our company with converting our SOPs and other controlled documents into electronic forms.  With this we will reduce the need to manually update physical SOPs as they can now be accessed through a web interface.